Move extractExplorerStateFromURL() to utils file
