+++ src/GuppyDataExplorer/utils.js
@@ -126,3 +126,38 @@ export function validateFilter(value, filterConfig, isAnchorFilterEnabled) {
 
   return allFields.length === testFieldSet.size;
 }
+
+/**
+ * @param {URLSearchParams} searchParams
+ * @param {FilterConfig} filterConfig
+ * @param {boolean} isAnchorFilterEnabled
+ * @param {PatientIdsConfig} [patientIdsConfig]
+ */
+export function extractExplorerStateFromURL(
+  searchParams,
+  filterConfig,
+  isAnchorFilterEnabled,
+  patientIdsConfig
+) {
+  /** @type {FilterState} */
+  let initialAppliedFilters = {};
+  if (searchParams.has('filter'))
+    try {
+      const filterInUrl = JSON.parse(decodeURI(searchParams.get('filter')));
+      if (validateFilter(filterInUrl, filterConfig, isAnchorFilterEnabled))
+        initialAppliedFilters = filterInUrl;
+      else throw new Error(undefined);
+    } catch (e) {
+      // eslint-disable-next-line no-console
+      console.error('Invalid filter value in URL.', e);
+    }
+
+  // eslint-disable-next-line no-nested-ternary
+  const patientIds = patientIdsConfig?.filter
+    ? searchParams.has('patientIds')
+      ? searchParams.get('patientIds').split(',')
+      : []
+    : undefined;
+
+  return { initialAppliedFilters, patientIds };
+}
